- Split builder, http listener, and push gateway into separate files of an exporter module
