+++ metrics-exporter-prometheus/src/lib.rs
@@ -106,8 +106,9 @@ pub use self::common::{BuildError, Matcher};
 mod distribution;
 pub use distribution::{Distribution, DistributionBuilder};
 
-mod builder;
-pub use self::builder::PrometheusBuilder;
+mod exporter;
+pub use self::exporter::builder::PrometheusBuilder;
+pub use self::exporter::ExporterFuture;
 
 pub mod formatting;
 mod recorder;
